Create quickbooksimportinvoiceform.cs
This is not modified for Laserfiche as the VM I had became corrupt.  This is the essentials for the code and needs maybe 10-20 minutes of variable changing to match LF system. It is set to fixed variables now.  Most of this code came from Quickbooks themselves and is fairly generic with a few modifications as this process has been implemented in a similar fashion to other systems many times before.